+++ Sandbox/GRCookbook/Bitmap.h
