+++ Sandbox/GRCookbook/Resource/GLFramebuffer.cpp
@@ -1,29 +1,7 @@
 ﻿#pragma once
 
-#include <glad/glad.h>
-#include <memory>
-
-class GLFramebuffer
-{
-public:
-	GLFramebuffer(int width, int height, GLenum formatColor, GLenum formatDepth);
-	~GLFramebuffer();
-	GLFramebuffer(const GLFramebuffer&) = delete;
-	GLFramebuffer(GLFramebuffer&&) = default;
-	GLuint getHandle() const { return handle_; }
-	const GLTexture& getTextureColor() const { return *texColor_.get(); }
-	const GLTexture& getTextureDepth() const { return *texDepth_.get(); }
-	void bind();
-	void unbind();
-
-private:
-	int width_;
-	int height_;
-	GLuint handle_ = 0;
-
-	std::unique_ptr<GLTexture> texColor_;
-	std::unique_ptr<GLTexture> texDepth_;
-};
+#include "GLFramebuffer.h"
+#include <cassert>
 
 GLFramebuffer::GLFramebuffer(int width, int height, GLenum formatColor, GLenum formatDepth)
 	: width_(width)
