+++ Sandbox/GRCookbook/Resource/GLSceneData.cpp
