+++ Sandbox/GRCookbook/Resource/GLSceneDataLazy.cpp
