+++ Sandbox/GRCookbook/UtilsCubemap.h
@@ -1,6 +1,6 @@
 ﻿#pragma once
 
-#include "shared/Bitmap.h"
+#include "Bitmap.h"
 
 Bitmap convertEquirectangularMapToVerticalCross(const Bitmap& b);
 Bitmap convertVerticalCrossToCubeMapFaces(const Bitmap& b);
