--- Sandbox/GRCookbook/shared/CMakeLists.txt
@@ -1,17 +1,0 @@
-cmake_minimum_required(VERSION 3.12)
-
-project(SharedUtils CXX C)
-
-include(../CMake/CommonMacros.txt)
-
-include_directories(../deps/src/imgui)
-
-file(GLOB_RECURSE SRC_FILES LIST_DIRECTORIES false RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} *.c??)
-file(GLOB_RECURSE HEADER_FILES LIST_DIRECTORIES false RELATIVE ${CMAKE_CURRENT_SOURCE_DIR} *.h)
-
-add_library(SharedUtils ${SRC_FILES} ${HEADER_FILES})
-
-set_property(TARGET SharedUtils PROPERTY CXX_STANDARD 20)
-set_property(TARGET SharedUtils PROPERTY CXX_STANDARD_REQUIRED ON)
-
-target_link_libraries(SharedUtils PUBLIC glad glfw)
