--- Sandbox/GRCookbook/shared/scene/MergeUtil.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "shared/scene/Scene.h"
-#include "shared/scene/VtxData.h"
-
-void mergeScene(Scene& scene, MeshData& meshData, const std::string& materialName);
